Developed main InvoiceService methods, DTOs and DTO builders.-

+++ src/main/java/com/coltraneinvoices/invoice/InvoiceDetail.java
@@ -39,7 +39,7 @@ public class InvoiceDetail {
 	private Invoice invoice;
 	
 	
-    private Integer productQuantity;
+    private Long productQuantity;
     
     private BigDecimal unitPrice;
